third_party/blink/web_tests/editing/shadow/selection-of-orphan-shadowroot-expected.txt

CONSOLE WARNING: line 16: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Nodes of the selection for an orphan shadow root should return null.

PASS selection.anchorNode is null
PASS selection.focusNode is null
PASS selection.baseNode is null
PASS selection.extentNode is null
PASS successfullyParsed is true

TEST COMPLETE
